refactor: extract video generation types and comfy client
